添加课程信息到服务器，从服务器获取课程信息

+++ app/src/main/java/com/chennyh/simpletimetable/bean/Course.java
@@ -9,9 +9,9 @@ import java.util.List;
 @Data
 public class Course implements ScheduleEnable {
 
-    private int id;
+    private Long id;
 
-    private int uid;
+    private Long uid;
 
     /**
      * 课程名
